+++ src/Config/Resilient.php
@@ -5,21 +5,32 @@ namespace Innmind\OperatingSystem\Config;
 
 use Innmind\OperatingSystem\Config;
 use Innmind\HttpTransport\Transport;
+use Innmind\Time\Period;
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
-enum Resilient
+final class Resilient
 {
-    case instance;
+    /**
+     * @param ?Sequence<Period> $retries
+     */
+    private function __construct(
+        private ?Sequence $retries,
+    ) {
+    }
 
     #[\NoDiscard]
     public function __invoke(Config $config): Config
     {
+        $retries = $this->retries;
+
         return $config
             ->mapHttpTransport(static fn($transport, $config) => Transport::exponentialBackoff(
                 $transport,
                 $config->halt(),
+                $retries,
             ));
     }
 
@@ -31,6 +42,15 @@ enum Resilient
     #[\NoDiscard]
     public static function new(): self
     {
-        return self::instance;
+        return new self(null);
+    }
+
+    /**
+     * @param Sequence<Period> $retries
+     */
+    #[\NoDiscard]
+    public function withHttpRetries(Sequence $retries): self
+    {
+        return new self($retries);
     }
 }